+++ src/allowNotifications.ts
@@ -1,5 +1,9 @@
 import bridge from '@vkontakte/vk-bridge';
 
+/*
+* Wrapper to VKWebAppAllowNotifications. Asks to allow notifications from app
+* @returns {boolean} Returns true if success
+*/
 export default async (): Promise<boolean> => {
   const { result } = await bridge.send('VKWebAppAllowNotifications');