+++ src/index.ts
@@ -1,31 +1,21 @@
 import addToFavorites from './addToFavorites';
 import allowMessagesFromGroup from './allowMessagesFromGroup';
 import allowNotifications from './allowNotifications';
-import getAuthToken, {
-  ScopesEnum,
-  AuthTokenResponseType,
-} from './getAuthToken';
+import getAuthToken from './getAuthToken';
 import getUserInfo from './getUserInfo';
 import setSwipeSettings from './setSwipeSettings';
-import setViewSettings, { ViewSettingsType } from './setViewSettings';
+import setViewSettings from './setViewSettings';
 import shareLink from './shareLink';
-import sharePost, {
-  AttachmentMediaEnum,
-  PostAttachmentType,
-} from './sharePost';
-import shareStory, {
-  BackgroundStoryEnum,
-  StoryAttachmentType,
-} from './shareStory';
+import sharePost from './sharePost';
+import shareStory from './shareStory';
 import {
   vibrateAsImpact,
   vibrateAsNotification,
   vibrateAsSelection,
-  VibrationImpactEnum,
-  VibrationNotificationEnum,
 } from './vibrate';
 import vkApi from './vkApi';
 import initializeVkApp from './initializeVkApp';
+import checkIOS from './checkIOS';
 
 export {
   addToFavorites,
@@ -43,16 +33,32 @@ export {
   vibrateAsSelection,
   vkApi,
   initializeVkApp,
+  checkIOS,
 };
 
-export {
-  ScopesEnum,
-  AuthTokenResponseType,
-  ViewSettingsType,
-  AttachmentMediaEnum,
+import {
+  PostAttachmentMediaEnum,
   PostAttachmentType,
   BackgroundStoryEnum,
   StoryAttachmentType,
+} from './types/sharing';
+import { ScopesEnum, AuthTokenResponseType } from './types/token';
+import {
   VibrationImpactEnum,
   VibrationNotificationEnum,
+} from './types/vabrations';
+import { ViewSettingsType } from './types/viewSettings';
+import { WindowType } from './types/window';
+
+export {
+  PostAttachmentMediaEnum,
+  PostAttachmentType,
+  BackgroundStoryEnum,
+  StoryAttachmentType,
+  ScopesEnum,
+  AuthTokenResponseType,
+  VibrationImpactEnum,
+  VibrationNotificationEnum,
+  ViewSettingsType,
+  WindowType,
 };