+++ src/setSwipeSettings.ts
@@ -1,5 +1,8 @@
 import bridge from '@vkontakte/vk-bridge';
 
+/*
+* Wrapper for VKWebAppSetSwipeSettings
+*/
 export default (history = true): void => {
   if (bridge.supports('VKWebAppSetSwipeSettings')) {
     bridge.send('VKWebAppSetSwipeSettings', { history });