+++ src/types/vabrations.ts
@@ -1,0 +1,11 @@
+export enum VibrationImpactEnum {
+    heavy = 'heavy',
+    medium = 'medium',
+    light = 'light',
+}
+
+export enum VibrationNotificationEnum {
+    error = 'error',
+    success = 'success',
+    warning = 'warning',
+}