+++ src/types/window.ts
@@ -1,0 +1,8 @@
+import { WindowType as CommonWindowType } from '@ktsstudio/mediaproject-utils';
+
+export interface WindowType extends CommonWindowType {
+  scope: string | null;
+  group_id: string | null;
+  page: string | null;
+  platform: string;
+}